tutorial, newsticker: land two agents' verification passes
Both agents went idle without reporting, so this lands their edits on the
strength of the annotations they left and my own reading of the parts that
matter most, not on a report. That is weaker provenance than the rest of
today's commits and is recorded as such; their findings should be re-read
before anything is built on them.

What I checked myself. -hitTest:withEvent: at 0x13c8a0 is untouched by the
agent and is correct as it stood: the guards return self while animating and on
the two terminal steps, and the spotlight test is fcmp d2, which is the clip
rect's width, matching what the source already tested. So the tutorial's own
hit-testing does not explain the overlay being stuck on the device, which
leaves the menu-button autoresizing mask fixed earlier as the live explanation
rather than a second fault here.

The news ticker's label was being given the base view's whole frame. It is
sized from the measured text instead, and only in width: the height is the one
the base view's frame left behind. That is the same shape as the two defects
found earlier today, a dimension taken from a register that an earlier call
left set, and it is a plausible cause of the blank ticker on the device,
though unproven until it is run.

The tutorial view gains constants for values that were inline, several with the
pool address they were decoded from, and four content-view dimensions become
single precision to match the properties they are assigned to.

+++ Project/RBMenuTutorialView.h
@@ -289,17 +289,22 @@ NS_ASSUME_NONNULL_BEGIN
 
 /**
  * @brief The message content view's width, derived from the iPad idiom.
+ *
+ * Single precision, not @c CGFloat: the accessors pass the value in @c s0 and every arithmetic use
+ * is a single-precision @c fmul widened afterwards with @c fcvt.
  * @ghidraAddress 0x1412f4 (getter)
  * @ghidraAddress 0x141304 (setter)
  */
-@property(nonatomic, assign) CGFloat contentViewWidth;
+@property(nonatomic, assign) float contentViewWidth;
 
 /**
  * @brief The message content view's height, derived from the iPad idiom.
+ *
+ * Single precision, not @c CGFloat, for the same reason as @c contentViewWidth.
  * @ghidraAddress 0x141314 (getter)
  * @ghidraAddress 0x141324 (setter)
  */
-@property(nonatomic, assign) CGFloat contentViewHeight;
+@property(nonatomic, assign) float contentViewHeight;
 
 @end
